fun with expect scripts - round 1
the grouper installer is very interactive, so fake it
with a long expect script. Note that you have to have
the mySQL database running for the install to work since
it wants to write to the database, and I hard coded a hostname
for the database server into the expect script for now.

It looks like there is a problem with the database not running with
the right character set on connect since there are conplaints in
the log in the /home/grouper/2.2.2/apache-tomcat-6.0.35/logs/grouperWs directory
like this:

Grouper error: grouper versions do not match, GrouperVersion.class: 2.2.2, grouper.version.properties: 2.2.2, manifest: null
Error: Cannot properly read UTF string from resource: grouperUtf8.txt: '???????', make sure you pass in the JVM switch -Dfile.encoding=utf-8 (currently is 'ANSI_X3.4-1968')

Signed-off-by: Mark McCahill <mark.mccahill@duke.edu>